WP-006: Fixing PostgreSQL image version (#6)
## 👨🏽‍💻 Description

This changes contain a quick fix to fix the version of the image for the
PostgreSQL container. The reason is because the previous one was
`postgis/postgis:14-3.3` which is specific for `x86_64`, the new one is
`1postgis/postgis:14-3.3` which also support `arm64`.

## ✅ Testing

I successfully built and ran the compose swarm within an `arm64` machine
and also in an `x86_64` machine. Also, verified the pipeline continue
passing.

## 🖼️ Evidence

Following is the screenshot of the build on `arm64`:


![image](https://github.com/zatarain/white-pages/assets/539783/5860c97e-735c-4b37-b4af-4fd648cc7100)

And in `x86_64`:


![image](https://github.com/zatarain/white-pages/assets/539783/b743cc4c-7fc0-4cea-b7f8-7921a9ffab07)

Signed-off-by: Ulises Tirado Zatarain <ulises.tirado@cimat.mx>